+++ test/test_temporal_interpolate.py
@@ -14,6 +14,7 @@ from test import TEST_DIR, WM
 import RAiDER
 from RAiDER.logger import logger
 
+
 def makeLatLonGrid(bbox, reg, out_dir, spacing=0.1):
     """ Make lat lons at a specified spacing """
     S, N, W, E = bbox
@@ -89,10 +90,10 @@ def test_cube_timemean():
         assert np.isclose(proc.returncode, 0)
 
     ## run interpolation in the middle of the two
-    grp['time_group'] =  {'time': ti}
+    grp['time_group'] =  {'time': ti, 'interpolate_wmtimes': True}
     cfg  = update_yaml(grp)
 
-    cmd  = f'raider.py ++process calcDelaysInterp {cfg}'
+    cmd  = f'raider.py {cfg}'
     proc = subprocess.run(cmd.split(), stdout=subprocess.PIPE, universal_newlines=True)
     assert np.isclose(proc.returncode, 0)
 
@@ -110,12 +111,13 @@ def test_cube_timemean():
     assert np.allclose(da_mu, da_interp_tot)
 
 
-    # # Clean up files
+    # Clean up files
     shutil.rmtree(SCENARIO_DIR)
     [os.remove(f) for f in glob.glob(f'{WM}*')]
     os.remove('temp.yaml')
 
     return
+
 
 def test_cube_weighting():
     """ Test the weighting by comparing a small crop with numpy directly """
@@ -126,7 +128,7 @@ def test_cube_weighting():
     S, N, W, E = 34, 35, -117, -116
     date       = 20200130
     dct_hrs    = {'GMAO': [12, 15, '12:05:00'], 'ERA5': [13, 14, '13:05:00']}
-    hr1, hr2, t_ref = dct_hrs[WM]
+    hr1, hr2, ti = dct_hrs[WM]
 
     grp = {
             'date_group': {'date_start': date},
@@ -148,10 +150,10 @@ def test_cube_weighting():
         assert np.isclose(proc.returncode, 0)
 
     ## run interpolation very near the first
-    grp['time_group'] =  {'time': t_ref}
+    grp['time_group'] =  {'time': ti, 'interpolate_wmtimes': True}
     cfg  = update_yaml(grp)
 
-    cmd  = f'raider.py ++process calcDelaysInterp {cfg}'
+    cmd  = f'raider.py {cfg}'
     proc = subprocess.run(cmd.split(), stdout=subprocess.PIPE, universal_newlines=True)
 
     ## double check on weighting
@@ -162,21 +164,30 @@ def test_cube_weighting():
     with xr.open_dataset(os.path.join(SCENARIO_DIR, f'{WM}_tropo_{date}T{hr2}0000_ztd.nc')) as ds:
         da2_tot = ds['wet'] + ds['hydro']
 
-    with xr.open_dataset(os.path.join(SCENARIO_DIR, f'{WM}_tropo_{date}T{t_ref.replace(":", "")}_ztd.nc')) as ds:
+    with xr.open_dataset(os.path.join(SCENARIO_DIR, f'{WM}_tropo_{date}T{ti.replace(":", "")}_ztd.nc')) as ds:
         da_interp_tot = ds['wet'] + ds['hydro']
 
     dt1 = datetime.strptime(f'{date}{hr1}', '%Y%m%d%H')
     dt2 = datetime.strptime(f'{date}{hr2}', '%Y%m%d%H')
-    dt_ref = datetime.strptime(f'{date}{t_ref}', '%Y%m%d%H:%M:%S')
+    dt_ref = datetime.strptime(f'{date}{ti}', '%Y%m%d%H:%M:%S')
 
     wgts  = np.array([(dt_ref-dt1).seconds, (dt2-dt_ref).seconds])
-    logger.info ('Weights: %s', wgts)
-    logger.info ('Tstart: %s, Tend: %s, Tref: %s', dt1, dt2, dt_ref)
     da1_crop = da1_tot.isel(z=0, y=slice(0,1), x=slice(0, 2))
     da2_crop = da2_tot.isel(z=0, y=slice(0,1), x=slice(0, 2))
     da_out_crop = da_interp_tot.isel(z=0, y=slice(0,1), x=slice(0,2))
 
     dat = np.vstack([da1_crop.data, da2_crop.data])
-    np.allclose(da_out_crop, np.average(dat, weights=1/wgts, axis=0))
+
+    logger.info ('Tstart: %s, Tend: %s, Tref: %s', dt1, dt2, dt_ref)
+    logger.info ('Weights: %s', wgts)
+    logger.info ('Data from two dates: %s', dat)
+    logger.info ('Weighted mean: %s', da_out_crop.data)
+    assert np.allclose(da_out_crop, np.average(dat, weights=1/wgts, axis=0))
+
+    # Clean up files
+    shutil.rmtree(SCENARIO_DIR)
+    [os.remove(f) for f in glob.glob(f'{WM}*')]
+    os.remove('temp.yaml')
+
     return